+++ lib/sheep_ast/messages.rb
@@ -20,6 +20,44 @@ module SheepAst
       @command = nil
       @description = nil
     end
+
+    def self.command_array(array)
+      arr = []
+      array.each do |elem|
+        arr << elem.command
+      end
+      return arr
+    end
+
+    def self.desc_array(array)
+      arr = []
+      array.each do |elem|
+        arr << elem.description
+      end
+      return arr
+    end
+  end
+
+  # TBD
+  class OnOff < T::Enum
+    enums do
+      Disable = new
+      Enable = new
+    end
+  end
+
+  # TBD
+  class OperateNode < T::Enum
+    include Log
+    include Exception
+
+    enums do
+      Top = new
+      Up = new
+      Goto = new
+      Revert = new
+      Commit = new
+    end
   end
 
   # Message struture between components
@@ -53,6 +91,20 @@ module SheepAst
       else
         application_error
       end
+    end
+  end
+
+  # Enum for action of matcher
+  #
+  # @api private
+  #
+  class MatchResult < T::Enum
+    enums do
+      NotFound = new
+      GetNext = new
+      Continue = new
+      Finish = new
+      Default = new
     end
   end
 
@@ -208,6 +260,7 @@ module SheepAst
   class AnalyzeData < T::Struct
     extend T::Sig
     prop :expr, T.nilable(String), default: nil
+    prop :is_eol, T.nilable(T::Boolean), default: nil
     prop :tokenized_line, T.nilable(T::Array[String]), default: nil
     prop :raw_line, T.nilable(String), default: nil
     prop :file_info, T.nilable(FileInfo), default: nil
@@ -220,6 +273,7 @@ module SheepAst
 
     def init
       @expr = nil
+      @is_eol = nil
       @tokenized_line = nil
       @file_info = nil
       @file_manager = nil
@@ -234,7 +288,8 @@ module SheepAst
     sig { returns(String) }
     def inspect
       return "custom inspect <#{self.class.name} object_id = #{object_id},"\
-        " expr = '#{expr.inspect}', stack = #{stack.inspect}, stack_symbol = #{stack_symbol.inspect},"\
+        " expr = '#{expr.inspect}', is_eol = '#{is_eol.inspect}',"\
+        " stack = #{stack.inspect}, stack_symbol = #{stack_symbol.inspect},"\
         " request_next_data = #{request_next_data.inspect}, file_info = #{file_info.inspect},"\
         " tokenized_line = #{@tokenized_line.inspect}, raw_line = #{@raw_line.inspect}"
     end